// src/modules/thread/enums/thread.enum.ts
export enum ThreadStatus {
    ACTIVE = 'active',
    ARCHIVED = 'archived',
    DELETED = 'deleted',
}

export enum ContentType {
    BLOG_POST = 'blog_post',
    PRODUCT_DESCRIPTION = 'product_description',
    SOCIAL_MEDIA_CAPTION = 'social_media_caption',
    ARTICLE = 'article',
    OTHER = 'other',
}